+++ android-json-form-wizard/src/test/java/com/vijay/jsonwizard/utils/TestClientForm.java
@@ -1,0 +1,111 @@
+package com.vijay.jsonwizard.utils;
+
+import org.smartregister.client.utils.contract.ClientFormContract;
+
+import java.util.Date;
+
+/**
+ * Created by Ephraim Kigamba - nek.eam@gmail.com on 01-07-2020.
+ */
+public class TestClientForm implements ClientFormContract.Model {
+    private int id;
+
+    private String version;
+
+    private String identifier;
+
+    private String module;
+
+    private String json;
+
+    private String jurisdiction;
+
+    private String label;
+
+    private boolean active;
+
+    private boolean isNew;
+
+    private Date createdAt;
+
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getVersion() {
+        return version;
+    }
+
+    public void setVersion(String version) {
+        this.version = version;
+    }
+
+    public String getIdentifier() {
+        return identifier;
+    }
+
+    public void setIdentifier(String identifier) {
+        this.identifier = identifier;
+    }
+
+    public String getModule() {
+        return module;
+    }
+
+    public void setModule(String module) {
+        this.module = module;
+    }
+
+    public String getJson() {
+        return json;
+    }
+
+    public void setJson(String json) {
+        this.json = json;
+    }
+
+    public String getJurisdiction() {
+        return jurisdiction;
+    }
+
+    public void setJurisdiction(String jurisdiction) {
+        this.jurisdiction = jurisdiction;
+    }
+
+    public String getLabel() {
+        return label;
+    }
+
+    public void setLabel(String label) {
+        this.label = label;
+    }
+
+    public boolean isActive() {
+        return active;
+    }
+
+    public void setActive(boolean active) {
+        this.active = active;
+    }
+
+    public boolean isNew() {
+        return isNew;
+    }
+
+    public void setNew(boolean aNew) {
+        isNew = aNew;
+    }
+
+    public Date getCreatedAt() {
+        return createdAt;
+    }
+
+    public void setCreatedAt(Date createdAt) {
+        this.createdAt = createdAt;
+    }
+}
